[Feature][User] User 쿼리폴더에 코드를 옮기고 상관 없는 코드 제거 했습니다.
