+++ src/main/java/com/seolandfriends/byeolbyeolcoffee/user/command/domain/aggregate/entity/User.java
@@ -7,11 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
-
-import java.util.List;
 
 
 @Entity
