+++ src/main/java/com/seolandfriends/byeolbyeolcoffee/user/query/domain/repository/UserQueryRepository.java
@@ -1,11 +1,10 @@
-package com.seolandfriends.byeolbyeolcoffee.user.command.domain.repository;
+package com.seolandfriends.byeolbyeolcoffee.user.query.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.seolandfriends.byeolbyeolcoffee.user.command.domain.aggregate.entity.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
+public interface UserQueryRepository extends JpaRepository<User, Integer> {
 	User findByUserId(Integer userId);
 
 	User findByUserAccount(String userAccount);
